+++ app/Http/Controllers/NivelFormacionController.php
@@ -5,69 +5,92 @@ namespace App\Http\Controllers;
 use App\Models\nivel_formacion;
 use App\Models\NivelFormacion;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+
+
 
 class NivelFormacionController extends Controller
 {
     /**
-     * Desplegar una lista de los recursos.
+     * Despliega una lista de recursos
+     * @return \Illuminate\Contracts\View\View
      */
     public function index()
     {
+        Gate::authorize('niveles_formacion.view', NivelFormacion::class);
         $niveles_formacion = NivelFormacion::all();
         return view('nivel_formacion.index', compact('niveles_formacion'));
     }
 
     /**
-     * Mostrar el formulario para crear un nuevo recurso.
+     * Despliega el formulario para crear un nuevo recurso
+     * @return \Illuminate\Contracts\View\View
      */
     public function create()
     {
+        Gate::authorize('niveles_formacion.create', NivelFormacion::class);
         return view('nivel_formacion.create');
     }
 
     /**
-     * Almacenar un recurso recién creado en almacenamiento.
+     * Crea un nuevo recurso
+     * @param Request $request
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function store(Request $request)
     {
+        Gate::authorize('niveles_formacion.create', NivelFormacion::class);
         NivelFormacion::create($request->all());
-        return redirect()->route('niveles_formacion.index');
+        return redirect()->route('niveles_formacion.index')->with('success', 'Nivel de formacion creado correctamente');
     }
 
     /**
-     * Desplegar el recurso especificado.
+     * Despliega el recurso especificado
+     * @param NivelFormacion $id
+     * @return \Illuminate\Contracts\View\View
      */
     public function show(NivelFormacion $id)
     {
+        Gate::authorize('niveles_formacion.view', $id);
         $nivel_formacion = NivelFormacion::find($id);
         return view('nivel_formacion.show', compact('nivel_formacion'));
     }
 
     /**
-     * Mostrar el formulario para editar el recurso especificado.
+     * Despliega el formulario para editar el recurso especificado
+     * @param NivelFormacion $id
+     * @return \Illuminate\Contracts\View\View
      */
     public function edit(NivelFormacion $id)
     {
+        Gate::authorize('niveles_formacion.update', $id);
         $nivel_formacion = NivelFormacion::find($id);
         return view('nivel_formacion.edit', compact('nivel_formacion'))->with('success', 'Nivel de formacion cargado correctamente para editar');
     }
 
     /**
-     * Actualizar el recurso especificado en almacenamiento.
+     * Actualiza el recurso especificado en almacenamiento
+     * @param Request $request
+     * @param NivelFormacion $id
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function update(Request $request, NivelFormacion $id)
     {
+        Gate::authorize('niveles_formacion.update', $id);
         $id->update($request->all());
-        return redirect()->route('nivel_formacion.index')->with('success', 'Nivel de formacion actualizado correctamente');
+        return redirect()->route('niveles_formacion.index')->with('success', 'Nivel de formacion actualizado correctamente');
     }
 
     /**
-     * Remover el recurso especificado de almacenamiento.
+     * Elimina el recurso especificado de almacenamiento
+     * @param NivelFormacion $id
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function destroy(NivelFormacion $id)
     {
+        Gate::authorize('niveles_formacion.delete', $id);
         $nivel_formacion = NivelFormacion::find($id);
         $nivel_formacion->delete();
-        return redirect()->route('nivel_formacion.index')->with('success', 'Nivel de formacion eliminado correctamente');
+        return redirect()->route('niveles_formacion.index')->with('success', 'Nivel de formacion eliminado correctamente');
     }
 }
